// src/lib/assets/types/nutrition.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Nutrition information for a food
 */
export interface Nutrition {
  /**
   * The number of servings of the food
   */
  servings: number;
  /**
   * The reccommended amount to eat per serving
   */
  servingSize: {
    /**
     * The numeric value for how much to eat
     */
    quantity: number;
    /**
     * The measurement of the quantity recommended to eat
     */
    unit: string;
    /**
     * The weight (in weightUnit, rounded up) recommended to consume
     */
    weight?: number;
    /**
     * The unit of weight/volume usu. g or ml
     */
    weightUnit?: string;
    [k: string]: unknown;
  };
  /**
   * The number of kilocalories per serving
   */
  calories: number;
  /**
   * The values for fats
   */
  fat: {
    /**
     * The total fat (in grams)
     */
    total: number;
    /**
     * The saturated fat (in grams)
     */
    saturated: number;
    /**
     * The trans fat (in grams)
     */
    trans: number;
    [k: string]: unknown;
  };
  /**
   * The value for cholesterol (in milligrams)
   */
  cholesterol: number;
  /**
   * The value for sodium (in milligrams)
   */
  sodium: number;
  /**
   * The values for carbohydrates (sugars)
   */
  carbs: {
    /**
     * The total carbs (in grams)
     */
    total: number;
    /**
     * The dietary fiber (in grams)
     */
    fiber: number;
    /**
     * The sugars
     */
    sugars: {
      /**
       * The total sugars (in grams)
       */
      total: number;
      /**
       * The added sugars (in grams)
       */
      added: number;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  /**
   * The value for protein (in grams)
   */
  protein: number;
  [k: string]: unknown;
}